refactor(exactly): ♻️ extract common code from ethereum
