+++ src/apps/exactly/ethereum/exactly.borrow.token-fetcher.ts
@@ -1,44 +1,6 @@
 import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
-import type { IMulticallWrapper } from '~multicall/multicall.interface';
-import type { AppTokenPosition } from '~position/position.interface';
-import type { GetDataPropsParams, GetTokenPropsParams } from '~position/template/app-token.template.types';
 
-import type { ExactlyMarketDefinition } from '../common/exactly.definitions-resolver';
-import { type ExactlyMarketProps, ExactlyTokenFetcher } from '../common/exactly.token-fetcher';
-import type { Market } from '../contracts';
+import { ExactlyBorrowFetcher } from '../common/exactly.borrow.token-fetcher';
 
 @PositionTemplate()
-export class EthereumExactlyBorrowFetcher extends ExactlyTokenFetcher {
-  groupLabel = 'Fixed Borrow';
-  isDebt = true;
-
-  getSupply({ definition }: GetTokenPropsParams<Market, ExactlyMarketProps, ExactlyMarketDefinition>) {
-    return Promise.resolve(definition.totalFloatingBorrowShares);
-  }
-
-  getTotalAssets({ definition }: GetTokenPropsParams<Market, ExactlyMarketProps, ExactlyMarketDefinition>) {
-    return definition.totalFloatingBorrowAssets;
-  }
-
-  getApr({ definition }: GetDataPropsParams<Market, ExactlyMarketProps, ExactlyMarketDefinition>) {
-    return Number(definition.floatingBorrowRate) / 1e16;
-  }
-
-  async getBalancePerToken({
-    address,
-    appToken,
-    multicall,
-  }: {
-    address: string;
-    appToken: AppTokenPosition;
-    multicall: IMulticallWrapper;
-  }) {
-    const { floatingBorrowShares } = await this.definitionsResolver.getDefinition({
-      multicall,
-      network: this.network,
-      account: address,
-      market: appToken.address,
-    });
-    return floatingBorrowShares;
-  }
-}
+export class EthereumExactlyBorrowFetcher extends ExactlyBorrowFetcher {}
